move ssr logic out of core
