+++ projects/core/public_api.ts
@@ -21,7 +21,6 @@ export * from './src/product/index';
 export * from './src/routing/index';
 export * from './src/site-context/index';
 export * from './src/smart-edit/index';
-export * from './src/ssr/index';
 export * from './src/state/index';
 export * from './src/store-finder/index';
 export * from './src/user/index';
