+++ projects/core/src/occ/config-loader/occ-config-loader.service.ts
@@ -16,7 +16,7 @@ import { map, tap } from 'rxjs/operators';
 import { deepMerge } from '../../config/utils/deep-merge';
 import { I18nConfig } from '../../i18n/config/i18n-config';
 import { SiteContextConfig } from '../../site-context/config/site-context-config';
-import { SERVER_REQUEST_URL } from '../../ssr/ssr.providers';
+import { SERVER_REQUEST_URL } from '../../util/ssr.tokens';
 import { OccLoadedConfig } from './occ-loaded-config';
 import { OccLoadedConfigConverter } from './occ-loaded-config-converter';
 import { OccSitesConfigLoader } from './occ-sites-config-loader';
