--- projects/core/src/ssr/index.ts
@@ -1,5 +1,0 @@
-export {
-  getServerRequestProviders,
-  NgExpressEngineDecorator,
-} from './ng-express-engine-decorator';
-export * from './ssr.providers';
