+++ projects/core/src/util/index.ts
@@ -9,3 +9,4 @@ export * from './withdraw-on';
 export * from './date-picker-formatter.service';
 export * from './date-time-picker-formatter.service';
 export * from './get-last-value-sync';
+export * from './ssr.tokens';
